A slightly simpler approach
This is a little more direct, but it requires that we put the code
that needs to be on both sides of the aisle into server, rather than
exporting code from the ingest-geoip module out into the license-aware
side of the house.

I'm keeping this here just as a WIP commit to give us a view of the
two sides at the same time -- we really could do it either way.
