+++ modules/ingest-geoip/src/main/java/module-info.java
@@ -16,5 +16,4 @@ module org.elasticsearch.ingest.geoip {
     requires com.maxmind.db;
 
     exports org.elasticsearch.ingest.geoip.stats to org.elasticsearch.server;
-    exports org.elasticsearch.ingest.geoip.enterprise to org.elasticsearch.geoipenterprisedownloader;
 }
