--- modules/ingest-geoip/src/main/java/org/elasticsearch/ingest/geoip/enterprise/EnterpriseGeoIpTaskParams.java
@@ -1,71 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-package org.elasticsearch.ingest.geoip.enterprise;
-
-import org.elasticsearch.TransportVersion;
-import org.elasticsearch.TransportVersions;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.persistent.PersistentTaskParams;
-import org.elasticsearch.xcontent.ObjectParser;
-import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xcontent.XContentParser;
-
-import java.io.IOException;
-
-import static org.elasticsearch.ingest.geoip.GeoIpDownloader.GEOIP_DOWNLOADER;
-
-public class EnterpriseGeoIpTaskParams implements PersistentTaskParams {
-
-    public static final String ENTERPRISE_GEOIP_DOWNLOADER = "enterprise-" + GEOIP_DOWNLOADER;
-
-    public static final ObjectParser<EnterpriseGeoIpTaskParams, Void> PARSER = new ObjectParser<>(
-        ENTERPRISE_GEOIP_DOWNLOADER,
-        true,
-        EnterpriseGeoIpTaskParams::new
-    );
-
-    public EnterpriseGeoIpTaskParams() {}
-
-    public EnterpriseGeoIpTaskParams(StreamInput in) {}
-
-    @Override
-    public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-        builder.startObject();
-        builder.endObject();
-        return builder;
-    }
-
-    @Override
-    public String getWriteableName() {
-        return ENTERPRISE_GEOIP_DOWNLOADER;
-    }
-
-    @Override
-    public TransportVersion getMinimalSupportedVersion() {
-        return TransportVersions.ENTERPRISE_GEOIP_DOWNLOADER;
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) {}
-
-    public static EnterpriseGeoIpTaskParams fromXContent(XContentParser parser) {
-        return PARSER.apply(parser, null);
-    }
-
-    @Override
-    public int hashCode() {
-        return 0;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return obj instanceof EnterpriseGeoIpTaskParams;
-    }
-}
